// src/swagger/model/userForListViewModel.ts
/**
 * NSI.Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface UserForListViewModel {
    userInfoId?: number;
    firstName?: string;
    lastName?: string;
    email?: string;
    isActive?: boolean;
    roles?: Array<Role>;
}